+++ packages/web-ui/src/index.ts
@@ -38,6 +38,6 @@ export * from './Text';
 export * from './TextField';
 export * from './TextLink';
 export * from './ThemeProvider';
-export * from './ToggleButton';
+export * from './ToggleButtonGroup';
 export * from './Typography';
 export * from './UnstyledButton';
